Fix editing an attachment' details, such as caption, for non-admins.
It removes the need for a user to be a member of the media site. Instead,
the user just needs to be able create attachments on the site itself.

We swap an `edit_posts` check for `create_posts` because the attachment
exists only on the network media site (it is not sync'd across sites).

Fixes #6