Ensure user can set a new featured image if image has been deleted.
Fixes #29

This applies the code from
https://github.com/humanmade/network-media-library/issues/29#issuecomment-434344263
to fix the issue where a media item has been deleted from the central
library but is referenced from another site.

It should be noted that this is masking over the problem a little which
is that when a media item is deleted for the site in which it's deleted
on it will also delete all the metadata links to that media item (see https://github.com/WordPress/WordPress/blob/master/wp-includes/post.php#L5425) this obviously doesn't do the same for all sites in the network. Therefore a more complete fix for this would be to hook into [delete_attachment](https://github.com/WordPress/WordPress/blob/master/wp-includes/post.php#L5419) looping over each site in the network and deleting the appropriate metadata.