Save the featured image via REST API if the media post ID doesn't exist in the non-media site
When using the Gutenberg editor to add a featured image, if the value of the post ID of the image on the media site doesn't exist in the non-media site that the featured image is being applied to, the featured image is not saved correctly.

This commit hooks into the `rest_insert_` action and saves the post ID of the featured image to the post meta table of the non-media site.