This commit will fix issue #51.
The issue is that the media site also needs to respect the
Advanced Custom Fields' return format. In the code we ask it to
kindly skip the formatting of the ACF if we are in the media site.
The fix is to allow the ACF to be formatted, independent of which
site or context the filter is run from.